+++ dejafu/Test/DejaFu/Internal.hs
@@ -41,7 +41,7 @@ data Settings n a = Settings
   , _debugPrint :: Maybe (String -> n ())
   , _earlyExit :: Maybe (Either Failure a -> Bool)
   , _equality :: Maybe (a -> a -> Bool)
-  , _shrink  :: Bool
+  , _simplify  :: Bool
   }
 
 -- | How to explore the possible executions of a concurrent program.